fix(core): correct edge-case handling in serialize, transport, and graph
Align int_or_any with its num/str/bool siblings so a present-but-wrong-typed
panMode/panLaw/channelDelaySamples falls back to the default instead of
aborting the whole project load; a numeric out-of-range value still errors.
Make the canonical scene loader reject out-of-range panMode/panLaw/negative
channelDelaySamples so the same JSON is accepted or refused identically on the
standalone and project-embedded paths.

Clamp bpm_at_sample to kMinBpm for constant segments so the reported tempo
matches the rate the playhead actually advances at, and fix the overstated
kMaxPublicPpq doc comment. Drop compile-derived state in Graph::remove_node so
node_latency_samples_q8 reports a removed node as unknown rather than a stale
value, and compute the port_data offset in size_t to avoid overflow for large
port and block sizes. Replace a raw default-BPM literal with constants::kDefaultBpm.

+++ src/graph/node.cpp
@@ -121,7 +121,7 @@ float* Node::port_data(std::vector<float>& storage, int port) noexcept {
   if (port < 0 || port >= num_ports_ || max_block_size_ <= 0) {
     return storage.data();
   }
-  return storage.data() + static_cast<size_t>(port * max_block_size_);
+  return storage.data() + static_cast<size_t>(port) * static_cast<size_t>(max_block_size_);
 }
 
 const float* Node::port_data(const std::vector<float>& storage, int port) const noexcept {
@@ -129,7 +129,7 @@ const float* Node::port_data(const std::vector<float>& storage, int port) const 
   if (port < 0 || port >= num_ports_ || max_block_size_ <= 0) {
     return storage.data();
   }
-  return storage.data() + static_cast<size_t>(port * max_block_size_);
+  return storage.data() + static_cast<size_t>(port) * static_cast<size_t>(max_block_size_);
 }
 
 }  // namespace sonare::graph
